refactor and pledge folder reorganisation
