+++ site/pages/pledge/[address].tsx
@@ -1,10 +1,12 @@
 import { GetStaticProps } from "next";
 
-import { PledgeDashboard } from "components/pages/Pledge/PledgeDashboard";
-import { pledgeResolver } from "lib/moralis";
-import { getPledgeByAddress } from "components/pages/Pledge/utils";
 import { loadTranslation } from "lib/i18n";
 import { IS_PRODUCTION } from "lib/constants";
+import { PledgeDashboard } from "components/pages/Pledge/PledgeDashboard";
+import {
+  getPledgeByAddress,
+  pledgeResolver,
+} from "components/pages/Pledge/lib";
 
 export const getStaticProps: GetStaticProps = async (ctx) => {
   if (IS_PRODUCTION) {
@@ -20,10 +22,9 @@ export const getStaticProps: GetStaticProps = async (ctx) => {
 
   try {
     const data = await getPledgeByAddress(address.toLowerCase());
-
     if (!data) throw new Error("Not found");
-    pledge = pledgeResolver(JSON.parse(JSON.stringify(data)));
-    console.log(pledge);
+
+    pledge = pledgeResolver(data);
   } catch (error) {
     pledge = null;
   }
